Many updates to get code working for epsilon < 1e-2.
Fixed implementation of constructor to actually use the prescribed initial guess (i.e. the diffusion initial guess).  Added a function to make a crude guess of the edge flux based on the initial average flux guess.

Updated definition of "error" to better account for overall convergence rate, including both CMFD and the SI part.

Updated what is returned by get_solution() to better reflect the latest data.

Add check that makes sure that the converged solution has all nonnegative fluxes.  If any fluxes are negative, it reports it as "unconverged."