Fix Qhat definition for reflective boundary, improve convergence criterion.
Qhat_edge is now set to 0 for a reflective boundary, which is correct since the 1st moment is trivially satisfied at all times for a reflective boundary.  No Qhat correction is needed.
Convergence criterion is now improved so that very small Qhat values (on the order of the tol) are ignored.